extracted bulk operations to a separate file
